Moves Bullet and some related classes/interfaces to a different file.
